backend/requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
blessed==1.20.0
cattrs==23.2.3
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.0.7
django-cors-headers==4.4.0
django-picklefield==3.2
django-q2==1.6.2
django-schema-graph==3.1.0
django-taggit==6.0.0
djangorestframework==3.15.2
graphqlclient==0.2.4
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
msgpack==1.0.8
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qrcode==7.4.2
redis==5.0.8
service-identity==24.1.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
wcwidth==0.2.13
zope.interface==7.0.1
